+++ CHANGELOG.md
@@ -8,8 +8,10 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ### Added
 
+- feature/1201611356837877 - Add callable collections to env file finder
 - feature/1201438240911768 - Combine ldl-env-util and ldl-file-finder-adapter-local to build compiled .env's from CLI
 - feature/1201475529448875 - Add more options to build command, fix directory exclusion bug
+
 
 ### Changed